requirements.txt
pandas
Pillow
pymongo
snscrape
streamlit
streamlit-option-menu
google-auth
imbalanced-learn
matplotlib
seaborn
numpy
scikit-learn
joblib
wordcloud